feature/todo-functionality
- refactor login/register functions
